Refactoring base AT
