+++ acceptance-tests/src/test/java/tech/pegasys/ethsigner/tests/AcceptanceTestBase.java
@@ -15,168 +15,60 @@ package tech.pegasys.ethsigner.tests;
 import static org.assertj.core.api.Assertions.assertThat;
 import static tech.pegasys.ethsigner.tests.WaitUtils.waitFor;
 
-import com.github.dockerjava.api.DockerClient;
-import com.github.dockerjava.api.command.CreateContainerCmd;
-import com.github.dockerjava.api.command.CreateContainerResponse;
-import com.github.dockerjava.api.command.DockerCmdExecFactory;
-import com.github.dockerjava.api.exception.NotFoundException;
-import com.github.dockerjava.api.exception.NotModifiedException;
-import com.github.dockerjava.api.model.HostConfig;
-import com.github.dockerjava.api.model.PortBinding;
-import com.github.dockerjava.core.DefaultDockerClientConfig;
-import com.github.dockerjava.core.DockerClientBuilder;
-import com.github.dockerjava.core.command.WaitContainerResultCallback;
-import com.github.dockerjava.jaxrs.JerseyDockerCmdExecFactory;
+import tech.pegasys.ethsigner.tests.dsl.Node;
+import tech.pegasys.ethsigner.tests.dsl.PantheonNode;
+import tech.pegasys.ethsigner.tests.dsl.Signer;
+
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
-import org.junit.After;
-import org.junit.Before;
+import org.junit.AfterClass;
+import org.junit.BeforeClass;
 import org.web3j.protocol.Web3j;
-import org.web3j.protocol.core.JsonRpc2_0Web3j;
-import org.web3j.protocol.http.HttpService;
-import org.web3j.utils.Async;
 
 public class AcceptanceTestBase {
 
   private static final Logger LOG = LogManager.getLogger();
-  private static final String LOCALHOST = "127.0.0.1";
 
-  private final Web3j ethSignerJsonRpc;
-  private final Web3j ethNodeJsonRpc;
+  private static Node ethNode;
+  private static Signer ehtSigner;
 
-  private EthSignerProcessRunner ethSignerRunner;
-  private DockerClient docker;
-  private String pantheonContainerId;
+  @BeforeClass
+  public static void setUpBase() {
+    Runtime.getRuntime().addShutdownHook(new Thread(AcceptanceTestBase::tearDownBase));
 
-  public AcceptanceTestBase() {
-    Runtime.getRuntime().addShutdownHook(new Thread(this::tearDownBase));
+    ehtSigner = new Signer();
+    ethNode = new PantheonNode();
 
-    ethSignerJsonRpc =
-        new JsonRpc2_0Web3j(
-            new HttpService("http://" + LOCALHOST + ":" + 9945),
-            2000,
-            Async.defaultExecutorService());
-
-    ethNodeJsonRpc =
-        new JsonRpc2_0Web3j(
-            new HttpService("http://" + LOCALHOST + ":" + 8545),
-            2000,
-            Async.defaultExecutorService());
-  }
-
-  @Before
-  public void setUpBase() {
-    docker = createDockerClient();
-    pantheonContainerId = createPantheonContainer();
-    startPantheonContainer();
-
-    ethSignerRunner = new EthSignerProcessRunner();
-    ethSignerRunner.start("EthSigner");
+    ethNode.start();
+    ehtSigner.start();
 
     awaitPantheonStartup();
     awaitEthSignerStartup();
   }
 
-  @After
-  public void tearDownBase() {
-    if (hasPantheonContainer()) {
-      stopPantheonContainer();
-      removePantheonContainer();
-    }
-
-    stopEthSigner();
+  @AfterClass
+  public static void tearDownBase() {
+    ethNode.shutdown();
+    ehtSigner.shutdown();
   }
 
   protected Web3j ethSigner() {
-    return ethSignerJsonRpc;
+    return ehtSigner.web3j();
   }
 
   protected Web3j ethNode() {
-    return ethNodeJsonRpc;
+    return ethNode.web3j();
   }
 
-  private void awaitEthSignerStartup() {
+  private static void awaitEthSignerStartup() {
     LOG.info("Waiting for EthSigner to become responsive...");
-    waitFor(() -> assertThat(ethSignerJsonRpc.ethBlockNumber().send().hasError()).isFalse());
+    waitFor(() -> assertThat(ehtSigner.web3j().ethBlockNumber().send().hasError()).isFalse());
     LOG.info("EthSigner is now responsive");
   }
 
-  private void awaitPantheonStartup() {
+  private static void awaitPantheonStartup() {
     LOG.info("Waiting for Pantheon to become responsive...");
-    waitFor(() -> assertThat(ethNodeJsonRpc.ethBlockNumber().send().hasError()).isFalse());
+    waitFor(() -> assertThat(ethNode.web3j().ethBlockNumber().send().hasError()).isFalse());
     LOG.info("Pantheon is now responsive");
-  }
-
-  private DockerClient createDockerClient() {
-    final DefaultDockerClientConfig config =
-        DefaultDockerClientConfig.createDefaultConfigBuilder().build();
-
-    final DockerCmdExecFactory dockerCmdExecFactory =
-        new JerseyDockerCmdExecFactory()
-            .withReadTimeout(1000)
-            .withConnectTimeout(1000)
-            .withMaxTotalConnections(100)
-            .withMaxPerRouteConnections(10);
-
-    return DockerClientBuilder.getInstance(config)
-        .withDockerCmdExecFactory(dockerCmdExecFactory)
-        .build();
-  }
-
-  private String createPantheonContainer() {
-    try {
-      // Bind the exposed 8545-8546 ports from the container to 8545-8546 on the host
-      final HostConfig portBindingConfig =
-          HostConfig.newHostConfig()
-              .withPortBindings(PortBinding.parse("8545:8545"), PortBinding.parse("8546:8546"));
-
-      final CreateContainerCmd createPantheon =
-          docker
-              .createContainerCmd("pegasyseng/pantheon:latest")
-              .withHostConfig(portBindingConfig)
-              .withCmd(
-                  "--miner-enabled",
-                  "--miner-coinbase",
-                  "fe3b557e8fb62b89f4916b721be55ceb828dbd73",
-                  "--rpc-http-cors-origins=\"all\"",
-                  "--rpc-http-enabled",
-                  "--rpc-ws-enabled",
-                  "--network=dev");
-
-      final CreateContainerResponse pantheon = createPantheon.exec();
-      return pantheon.getId();
-    } catch (final NotFoundException e) {
-      throw new RuntimeException(
-          "Before you run the acceptance tests, execute 'docker pull pegasyseng/pantheon:latest'",
-          e);
-    }
-  }
-
-  private void startPantheonContainer() {
-    docker.startContainerCmd(pantheonContainerId).exec();
-  }
-
-  private boolean hasPantheonContainer() {
-    return docker != null && pantheonContainerId != null;
-  }
-
-  private void stopPantheonContainer() {
-    try {
-      LOG.info("Stopping the Pantheon Docker container");
-      docker.stopContainerCmd(pantheonContainerId).exec();
-      docker.waitContainerCmd(pantheonContainerId).exec((new WaitContainerResultCallback()));
-    } catch (final NotModifiedException e) {
-      LOG.error("Pantheon Docker container has already stopped");
-    }
-  }
-
-  private void removePantheonContainer() {
-    LOG.info("Removing the Pantheon Docker container");
-    docker.removeContainerCmd(pantheonContainerId).withForce(true).exec();
-  }
-
-  private void stopEthSigner() {
-    LOG.info("Shutting down EthSigner");
-    ethSignerRunner.shutdown();
   }
 }
